fix: Option deletion and class name change
