--- Assets/KoeiromapUnity/Scripts/Option.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: dde29c0c8e014ef3850348fc7f3e1bb9
-timeCreated: 1679329502
